refactor: connect to known peers
