+++ src/common/constants.ts
@@ -1,0 +1,1 @@
+export const FIBER_RPC_URL = process.env["FIBER_RPC_URL"] ?? "http://localhost:58227";
